fx sorting & drop currency list

+++ app/modules/WalletList/helpers.js
@@ -225,7 +225,7 @@ const getSortedData = (array, currentArray, accountList, filter) => {
         }
         case 'byValue': {
             let sortedAccount = _orderBy(accountList, function(obj) {
-                return parseInt(obj.basicCurrencyBalance.toString().replace(/\s+/g, ''), 10)
+                return parseFloat(obj.basicCurrencyBalance.toString().replace(/\s+/g, ''), 10)
             }, 'desc').map(item => item.currencyCode)
 
             sortedAccount = _orderBy(array, x => {
